ajout du nom du fichier uploadé dans l'input et changement du integer…

+++ src/Entity/Ad.php
@@ -73,7 +73,7 @@ class Ad
     private $type;
 
     /**
-     * @ORM\Column(type="integer", nullable=true)
+     * @ORM\Column(type="float", nullable=true)
      */
     private $price;
 
@@ -228,12 +228,12 @@ class Ad
         return $this;
     }
 
-    public function getPrice(): ?int
+    public function getPrice(): ?float
     {
         return $this->price;
     }
 
-    public function setPrice(?int $price): self
+    public function setPrice(?float $price): self
     {
         $this->price = $price;
